Changes Updation

+++ src/main/java/com/assessment/service/impl/EmployeeServiceImpl.java
@@ -1,14 +1,14 @@
-package com.assessment.serviceImpl;
+package com.assessment.service.impl;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-
-import javax.validation.Valid;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
 import com.assessment.exception.UserAlredyExist;
+import com.assessment.exception.UserNotExist;
 import com.assessment.model.Employee;
 import com.assessment.model.EmployeeModel;
 import com.assessment.repository.EmployeeRepository;
@@ -25,9 +25,8 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	@Override
 	public List<EmployeeModel> getAllEmployees() {
-		List<Employee> employees = new ArrayList<Employee>();
-		employeeRepository.findAll().forEach((employee) -> employees.add(employee));
-		List<EmployeeModel> employeesModel = new ArrayList<EmployeeModel>();
+		List<Employee> employees = employeeRepository.findAll();
+		List<EmployeeModel> employeesModel = new ArrayList<>();
 
 		for (Employee e : employees) {
 			EmployeeModel em = new EmployeeModel();
@@ -45,8 +44,11 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	@Override
 	public EmployeeModel getEmployeeById(int id) {
-		Employee employee = employeeRepository.findById(id).get();
+		Optional<Employee> e = employeeRepository.findById(id);
+		if (!e.isPresent())
+			throw new UserNotExist("User doesn't exist with id : " + id);
 
+		Employee employee = e.get();
 		EmployeeModel em = new EmployeeModel();
 		em.setId(employee.getId());
 		em.setDesignation(employee.getDesignation());
@@ -59,10 +61,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 	}
 
 	@Override
-	public EmployeeModel createEmployee(@Valid EmployeeModel employeeModel) {
-		Employee e  = employeeRepository.findFirstByEmailId(employeeModel.getEmail());
-		if(Objects.nonNull(e)) {
-			throw new UserAlredyExist("User Alredy Exist with email : "+e.getEmail());
+	public boolean createEmployee(EmployeeModel employeeModel) {
+		Employee e = employeeRepository.findFirstByEmailId(employeeModel.getEmail());
+		if (Objects.nonNull(e)) {
+			throw new UserAlredyExist("User Alredy Exist with email : " + e.getEmail());
 		}
 		Employee employee = new Employee();
 		employee.setName(employeeModel.getName());
@@ -71,32 +73,39 @@ public class EmployeeServiceImpl implements EmployeeService {
 		employee.setPhone(employeeModel.getPhone());
 		employee.setSalary(employeeModel.getSalary());
 		employee.setTech(employeeModel.getTech());
-		employeeRepository.save(employee);
-		return employeeModel;
-
+		try {
+			employeeRepository.save(employee);
+			return true;
+		} catch (Exception ex) {
+			return false;
+		}
 	}
 
 	@Override
-	public EmployeeModel updateEmployee(int id, EmployeeModel employeeDetails) {
-		Employee updateEmployee = null;
-
-		updateEmployee = employeeRepository.findById(id).get();
-
-		updateEmployee.setName(employeeDetails.getName());
-		updateEmployee.setEmail(employeeDetails.getEmail());
-		updateEmployee.setDesignation(employeeDetails.getDesignation());
-		updateEmployee.setPhone(employeeDetails.getPhone());
-		updateEmployee.setSalary(employeeDetails.getSalary());
-		updateEmployee.setTech(employeeDetails.getTech());
-
-		employeeRepository.save(updateEmployee);
-		return employeeDetails;
+	public boolean updateEmployee(EmployeeModel employeeDetails) {
+		Optional<Employee> updateEmployee = employeeRepository.findById(employeeDetails.getId());
+		Employee updation = null;
+		if (updateEmployee.isPresent()) {
+			updation = updateEmployee.get();
+			updation.setName(employeeDetails.getName());
+			updation.setEmail(employeeDetails.getEmail());
+			updation.setDesignation(employeeDetails.getDesignation());
+			updation.setPhone(employeeDetails.getPhone());
+			updation.setSalary(employeeDetails.getSalary());
+			updation.setTech(employeeDetails.getTech());
+			employeeRepository.save(updation);
+			return true;
+		}
+		return false;
 	}
 
 	@Override
-	public int deleteEmployeeById(int id) {
-		employeeRepository.deleteById(id);
-		return id;
+	public boolean deleteEmployeeById(int id) {
+		if (employeeRepository.findById(id).isPresent()) {
+			employeeRepository.deleteById(id);
+			return true;
+		}
+		throw new UserNotExist("Unable to perform delete as User doesn't exist with id : " + id);
 
 	}
